requirements.txt
numpy
matplotlib
numba
pillow
imageio
psutil
pynvml
